+++ src/main/java/org/project/currencyexchange/service/ExternalExchangeApi.java
@@ -1,7 +1,8 @@
 package org.project.currencyexchange.service;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 public interface ExternalExchangeApi {
-    Map<String, Double> getExchangeRates(String baseCurrency);
+    Map<String, BigDecimal> getExchangeRates(String baseCurrency);
 }
